+++ cmd/function_deploy.go
@@ -5,11 +5,11 @@ package cmd
 import (
 	"fmt"
 	"github.com/vertigobr/safira/pkg/get"
-	"github.com/vertigobr/safira/pkg/stack"
+	s "github.com/vertigobr/safira/pkg/stack"
 	"os"
 
 	"github.com/vertigobr/safira/pkg/config"
-	"github.com/vertigobr/safira/pkg/deploy"
+	d "github.com/vertigobr/safira/pkg/deploy"
 	"github.com/vertigobr/safira/pkg/execute"
 
 	"github.com/spf13/cobra"
@@ -59,18 +59,20 @@ func runFunctionDeploy(cmd *cobra.Command, args []string) error {
 	}
 
 	kubectlPath := config.GetKubectlPath()
-	functions, err := stack.GetAllFunctions()
+	stack, err := s.LoadStackFile()
 	if err != nil {
 		return err
 	}
 
+	functions := stack.Functions
 	if all {
 		for index, _ := range functions {
 			if err := checkDeployFiles(index, hostnameFlag); err!= nil {
 				return err
 			}
 
-			if err := functionDeploy(kubectlPath, kubeconfigFlag, index, verboseFlag, updateFlag); err != nil {
+			deployFolder := fmt.Sprintf("deploy/%s/", index)
+			if err := deploy(kubectlPath, kubeconfigFlag, deployFolder, index, verboseFlag, updateFlag); err != nil {
 				return err
 			}
 		}
@@ -81,11 +83,20 @@ func runFunctionDeploy(cmd *cobra.Command, args []string) error {
 					return err
 				}
 
-				if err := functionDeploy(kubectlPath, kubeconfigFlag, functionArg, verboseFlag, updateFlag); err != nil {
+				deployFolder := fmt.Sprintf("deploy/%s/", functionArg)
+				if err := deploy(kubectlPath, kubeconfigFlag, deployFolder, functionArg, verboseFlag, updateFlag); err != nil {
 					return err
 				}
 			} else {
 				return fmt.Errorf("nome dá função %s é inválido", functionArg)
+			}
+		}
+	}
+
+	if len(stack.Custom) > 0 {
+		for _, path := range stack.Custom {
+			if err := deploy(kubectlPath, kubeconfigFlag, path, "", verboseFlag, updateFlag); err != nil {
+				return err
 			}
 		}
 	}
@@ -95,7 +106,7 @@ func runFunctionDeploy(cmd *cobra.Command, args []string) error {
 	return nil
 }
 
-func functionDeploy(kubectlPath, kubeconfigFlag, functionName string, verboseFlag, updateFlag bool) error {
+func deploy(kubectlPath, kubeconfigFlag, deployFolder, functionName string, verboseFlag, updateFlag bool) error {
 	var kubeconfig string
 	if len(kubeconfigFlag) > 0 {
 		kubeconfig = kubeconfigFlag
@@ -108,39 +119,20 @@ func functionDeploy(kubectlPath, kubeconfigFlag, functionName string, verboseFla
 		kubeconfig = config.GetKubeconfig()
 	}
 
-	hasFunction, err := deploy.CheckFunction(clusterName, functionName, functionsNamespace)
-	if err != nil {
-		return err
-	}
-
-	if hasFunction && updateFlag {
-		taskRemoveFunction := execute.Task{
-			Command:     kubectlPath,
-			Args:        []string{
-				"rollout", "restart", "deployments", functionName,
-				"-n", functionsNamespace,
-				"--kubeconfig", kubeconfig,
-			},
-			StreamStdio:  verboseFlag,
-			PrintCommand: verboseFlag,
-		}
-
-		if verboseFlag {
-			fmt.Printf("[+] Reiniciando a função " + functionName)
-		}
-
-		res, err := taskRemoveFunction.Execute()
+	if len(functionName) > 0 {
+		hasFunction, err := d.CheckFunction(clusterName, functionName, functionsNamespace)
 		if err != nil {
 			return err
 		}
 
-		if res.ExitCode != 0 {
-			return fmt.Errorf(res.Stderr)
+		if hasFunction && updateFlag {
+			if err := rolloutFunction(kubectlPath, kubeconfig, functionName, verboseFlag); err != nil {
+				return err
+			}
 		}
 	}
 
-	deployFolder := fmt.Sprintf("deploy/%s/", functionName)
-	taskFunctionDeploy := execute.Task{
+	taskDeploy := execute.Task{
 		Command:     kubectlPath,
 		Args:        []string{
 			"apply", "--wait",
@@ -151,8 +143,13 @@ func functionDeploy(kubectlPath, kubeconfigFlag, functionName string, verboseFla
 		PrintCommand: verboseFlag,
 	}
 
-	fmt.Println("Executando deploy da função " + functionName + "...")
-	res, err := taskFunctionDeploy.Execute()
+	if len(functionName) > 0 {
+		fmt.Println("Executando deploy da função " + functionName + "...")
+	} else {
+		fmt.Println("Executando deploy de arquivos customizados, " + deployFolder)
+	}
+
+	res, err := taskDeploy.Execute()
 	if err != nil {
 		return err
 	}
@@ -176,16 +173,44 @@ func checkDeployFiles(functionName, hostnameFlag string) error {
 		}
 	}
 
-	if err := deploy.CreateYamlFunction(functionYaml, functionName, functionsNamespace); err != nil {
+	if err := d.CreateYamlFunction(functionYaml, functionName, functionsNamespace); err != nil {
 		return err
 	}
 
-	if err := deploy.CreateYamlIngress(ingressYaml, functionName, hostnameFlag); err != nil {
+	if err := d.CreateYamlIngress(ingressYaml, functionName, hostnameFlag); err != nil {
 		return err
 	}
 
-	if err := deploy.CreateYamlService(serviceYaml, functionName, hostnameFlag); err != nil {
+	if err := d.CreateYamlService(serviceYaml, functionName, hostnameFlag); err != nil {
 		return err
+	}
+
+	return nil
+}
+
+func rolloutFunction(kubectlPath, kubeconfig, functionName string, verboseFlag bool) error {
+	taskRemoveFunction := execute.Task{
+		Command:     kubectlPath,
+		Args:        []string{
+			"rollout", "restart", "deployments", functionName,
+			"-n", functionsNamespace,
+			"--kubeconfig", kubeconfig,
+		},
+		StreamStdio:  verboseFlag,
+		PrintCommand: verboseFlag,
+	}
+
+	if verboseFlag {
+		fmt.Println("[+] Reiniciando a função " + functionName)
+	}
+
+	res, err := taskRemoveFunction.Execute()
+	if err != nil {
+		return err
+	}
+
+	if res.ExitCode != 0 {
+		return fmt.Errorf(res.Stderr)
 	}
 
 	return nil